backend/policeReports/requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-admin==2.0.1
django-cors-headers==3.4.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.11.0
docker==2.0.0
docker-pycreds==0.4.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.5.3
pytz==2020.1
requests==2.24.0
screen==1.0.1
six==1.12.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
websocket-client==0.57.0
wrapt==1.12.1
xlwt==1.3.0
